requirements.txt
#tensorflow version
# tensorflow
# numpy
# matplotlib

# pytorch version
torch
torchvision
torchaudio
matplotlib
Pillow
scikit-learn
seaborn
wandb